tests/requirements.txt
pytest>=8.0
pytest-asyncio>=0.24
httpx>=0.28.0
sqlalchemy[asyncio]>=2.0.36
psycopg[binary]>=3.2.0
pgvector>=0.3.6
pydantic-settings>=2.6.0
pypdf>=5.0.0
python-docx>=1.1.0
structlog>=24.4.0
reportlab>=4.0
watchdog>=6.0.0
mcp[cli]>=1.2.0
